Fix typing and clarify board indexes
I fixed the situations in the code where it switched types of variables
from numbers to booleans, so that the variables now stay in their type.
Also, I made the renamed the board indexes from "a" to "blank", from "b"
to "computer", and from "c" to "user". This makes the code more
readable.